+++ RELEASES.md
@@ -4,9 +4,9 @@
 
 ### Breaking Changes
 
-- removed `block_ui_interactions`:
-  - by default, this library will prioritize `bevy::ui`, thus you could simply remove the `block_ui_interactions` feature from your `Cargo.toml`.
-  - if you want to disable this priority, add the `no_ui_priority` feature to your configuration.
+- removed the `block_ui_interactions` feature:
+  - by default, this library will prioritize `bevy::ui`.
+  - if you want to disable this priority, add the newly added `no_ui_priority` feature to your configuration.
 
 ## Version 0.13.0